change folders and added index imports
