+++ apps/app/src/data/domain/list/Screen.ts
@@ -1,0 +1,7 @@
+export enum Screen {
+  Home = "Home",
+  Crew = "Crew",
+  Quests = "Quests",
+  CreateTrip = "CreateTrip",
+  CreateQuest = "CreateQuest",
+}
