+++ apps/app/src/data/domain/list/Trips.ts
