+++ apps/app/src/types/domain/Trip.ts
