+++ apps/app/src/types/generic/Timer.ts
